add more extensive wasm recompilation tests

+++ system_tests/program_test.go
@@ -50,6 +50,8 @@ var oneEth = arbmath.UintToBig(1e18)
 
 var allWasmTargets = []string{string(rawdb.TargetWavm), string(rawdb.TargetArm64), string(rawdb.TargetAmd64), string(rawdb.TargetHost)}
 
+var localTargetOnly = []string{string(rawdb.LocalTarget())}
+
 func TestProgramKeccak(t *testing.T) {
 	t.Parallel()
 	t.Run("WithDefaultWasmTargets", func(t *testing.T) {
@@ -64,7 +66,7 @@ func TestProgramKeccak(t *testing.T) {
 
 	t.Run("WithOnlyLocalTarget", func(t *testing.T) {
 		keccakTest(t, true, func(builder *NodeBuilder) {
-			builder.WithExtraArchs([]string{string(rawdb.LocalTarget())})
+			builder.WithExtraArchs(localTargetOnly)
 		})
 	})
 }
@@ -1938,7 +1940,7 @@ func formatTime(duration time.Duration) string {
 	return fmt.Sprintf("%.2f%s", span, units[unit])
 }
 
-func testWasmRecreate(t *testing.T, builder *NodeBuilder, storeTx *types.Transaction, loadTx *types.Transaction, want []byte) {
+func testWasmRecreate(t *testing.T, builder *NodeBuilder, targetsBefore, targetsAfter []string, numModules int, removeWasmDbBetween bool, storeTx, loadTx *types.Transaction, want []byte) {
 	ctx := builder.ctx
 	l2info := builder.L2Info
 	l2client := builder.L2.Client
@@ -1950,7 +1952,9 @@ func testWasmRecreate(t *testing.T, builder *NodeBuilder, storeTx *types.Transac
 
 	testDir := t.TempDir()
 	nodeBStack := testhelpers.CreateStackConfigForTest(testDir)
-	nodeB, cleanupB := builder.Build2ndNode(t, &SecondNodeParams{stackConfig: nodeBStack})
+	nodeBExecConfigBefore := *builder.execConfig
+	nodeBExecConfigBefore.StylusTarget.ExtraArchs = targetsBefore
+	nodeB, cleanupB := builder.Build2ndNode(t, &SecondNodeParams{stackConfig: nodeBStack, execConfig: &nodeBExecConfigBefore})
 
 	_, err = EnsureTxSucceeded(ctx, nodeB.Client, storeTx)
 	Require(t, err)
@@ -1961,21 +1965,26 @@ func testWasmRecreate(t *testing.T, builder *NodeBuilder, storeTx *types.Transac
 	if !bytes.Equal(result, want) {
 		t.Fatalf("got wrong value, got %x, want %x", result, want)
 	}
+	wasmDb := nodeB.ExecNode.Backend.ArbInterface().BlockChain().StateCache().WasmStore()
+	checkWasmStoreContent(t, wasmDb, nodeBExecConfigBefore.StylusTarget.WasmTargets(), numModules)
 	// close nodeB
 	cleanupB()
 
-	// delete wasm dir of nodeB
-
-	wasmPath := filepath.Join(testDir, "system_tests.test", "wasm")
-	dirContents, err := os.ReadDir(wasmPath)
-	Require(t, err)
-	if len(dirContents) == 0 {
-		Fatal(t, "not contents found before delete")
+	wasmPath := filepath.Join(testDir, nodeBStack.Name, "wasm")
+	if removeWasmDbBetween {
+		// remove wasm dir of nodeB
+		dirContents, err := os.ReadDir(wasmPath)
+		Require(t, err)
+		if len(dirContents) == 0 {
+			Fatal(t, "not contents found before delete")
+		}
+		os.RemoveAll(wasmPath)
 	}
-	os.RemoveAll(wasmPath)
 
 	// recreate nodeB - using same source dir (wasm deleted)
-	nodeB, cleanupB = builder.Build2ndNode(t, &SecondNodeParams{stackConfig: nodeBStack})
+	nodeBExecConfigAfter := *builder.execConfig
+	nodeBExecConfigAfter.StylusTarget.ExtraArchs = targetsAfter
+	nodeB, cleanupB = builder.Build2ndNode(t, &SecondNodeParams{stackConfig: nodeBStack, execConfig: &nodeBExecConfigAfter})
 
 	// test nodeB - sees existing transaction
 	_, err = EnsureTxSucceeded(ctx, nodeB.Client, storeTx)
@@ -1988,6 +1997,9 @@ func testWasmRecreate(t *testing.T, builder *NodeBuilder, storeTx *types.Transac
 		t.Fatalf("got wrong value, got %x, want %x", result, want)
 	}
 
+	//wasmDb = nodeB.ExecNode.Backend.ArbInterface().BlockChain().StateCache().WasmStore()
+	//checkWasmStoreContent(t, wasmDb, nodeBExecConfigAfter.StylusTarget.WasmTargets(), numModules)
+
 	// send new tx (requires wasm) and check nodeB sees it as well
 	Require(t, l2client.SendTransaction(ctx, loadTx))
 
@@ -1997,16 +2009,85 @@ func testWasmRecreate(t *testing.T, builder *NodeBuilder, storeTx *types.Transac
 	_, err = EnsureTxSucceeded(ctx, nodeB.Client, loadTx)
 	Require(t, err)
 
+	wasmDb = nodeB.ExecNode.Backend.ArbInterface().BlockChain().StateCache().WasmStore()
+	checkWasmStoreContent(t, wasmDb, nodeBExecConfigAfter.StylusTarget.WasmTargets(), numModules)
+
 	cleanupB()
-	dirContents, err = os.ReadDir(wasmPath)
+	dirContents, err := os.ReadDir(wasmPath)
 	Require(t, err)
 	if len(dirContents) == 0 {
-		Fatal(t, "not contents found before delete")
+		Fatal(t, "no contents found before delete")
 	}
 	os.RemoveAll(wasmPath)
 }
 
 func TestWasmRecreate(t *testing.T) {
+	testCases := []struct {
+		name                string
+		removeWasmDbBetween bool
+		targetsBefore       []string
+		targetsAfter        []string
+	}{
+		{
+			name:                "with local target only with wasmdb removal",
+			removeWasmDbBetween: true,
+			targetsBefore:       localTargetOnly,
+			targetsAfter:        localTargetOnly,
+		},
+		{
+			name:                "with local target only without wasmdb removal",
+			removeWasmDbBetween: false,
+			targetsBefore:       localTargetOnly,
+			targetsAfter:        localTargetOnly,
+		},
+		{
+			name:                "with all targets with wasmdb removal",
+			removeWasmDbBetween: true,
+			targetsBefore:       allWasmTargets,
+			targetsAfter:        allWasmTargets,
+		},
+		{
+			name:                "with all targets without wasmdb removal",
+			removeWasmDbBetween: false,
+			targetsBefore:       allWasmTargets,
+			targetsAfter:        allWasmTargets,
+		},
+		{
+			name:                "more targets to recreate with wasmdb removal",
+			removeWasmDbBetween: true,
+			targetsBefore:       localTargetOnly,
+			targetsAfter:        allWasmTargets,
+		},
+		{
+			name:                "more targets to recreate without wasmdb removal",
+			removeWasmDbBetween: false,
+			targetsBefore:       localTargetOnly,
+			targetsAfter:        allWasmTargets,
+		},
+		{
+			name:                "less targets to recreate with wasmdb removal",
+			removeWasmDbBetween: true,
+			targetsBefore:       allWasmTargets,
+			targetsAfter:        localTargetOnly,
+		},
+		{
+			name:                "less targets to recreate without wasmdb removal",
+			removeWasmDbBetween: false,
+			targetsBefore:       allWasmTargets,
+			targetsAfter:        localTargetOnly,
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			testWasmRecreateWithCall(t, tc.targetsBefore, tc.targetsAfter, tc.removeWasmDbBetween)
+		})
+		t.Run(tc.name+" with delegate call", func(t *testing.T) {
+			testWasmRecreateWithDelegatecall(t, tc.targetsBefore, tc.targetsAfter, tc.removeWasmDbBetween)
+		})
+	}
+}
+
+func testWasmRecreateWithCall(t *testing.T, targetsBefore, targetsAfter []string, removeWasmDbBetween bool) {
 	builder, auth, cleanup := setupProgramTest(t, true)
 	ctx := builder.ctx
 	l2info := builder.L2Info
@@ -2021,10 +2102,10 @@ func TestWasmRecreate(t *testing.T) {
 	storeTx := l2info.PrepareTxTo("Owner", &storage, l2info.TransferGas, nil, argsForStorageWrite(zero, val))
 	loadTx := l2info.PrepareTxTo("Owner", &storage, l2info.TransferGas, nil, argsForStorageRead(zero))
 
-	testWasmRecreate(t, builder, storeTx, loadTx, val[:])
+	testWasmRecreate(t, builder, localTargetOnly, allWasmTargets, 1, false, storeTx, loadTx, val[:])
 }
 
-func TestWasmRecreateWithDelegatecall(t *testing.T) {
+func testWasmRecreateWithDelegatecall(t *testing.T, targetsBefore, targetsAfter []string, removeWasmDbBetween bool) {
 	builder, auth, cleanup := setupProgramTest(t, true)
 	ctx := builder.ctx
 	l2info := builder.L2Info
@@ -2043,7 +2124,7 @@ func TestWasmRecreateWithDelegatecall(t *testing.T) {
 	data = argsForMulticall(vm.DELEGATECALL, storage, big.NewInt(0), argsForStorageRead(zero))
 	loadTx := l2info.PrepareTxTo("Owner", &multicall, l2info.TransferGas, nil, data)
 
-	testWasmRecreate(t, builder, storeTx, loadTx, val[:])
+	testWasmRecreate(t, builder, localTargetOnly, allWasmTargets, 2, true, storeTx, loadTx, val[:])
 }
 
 // createMapFromDb is used in verifying if wasm store rebuilding works
@@ -2112,7 +2193,7 @@ func TestWasmStoreRebuilding(t *testing.T) {
 	cleanupB()
 
 	// delete wasm dir of nodeB
-	wasmPath := filepath.Join(testDir, "system_tests.test", "wasm")
+	wasmPath := filepath.Join(testDir, nodeBStack.Name, "wasm")
 	dirContents, err := os.ReadDir(wasmPath)
 	Require(t, err)
 	if len(dirContents) == 0 {
